factor out subteams list
